Enable session resume and signal strength check
Once a socket has been established and we have previously returned
from a successful handshake we should reuse the agreement and resume
the previous TLS session for all subsequent sessions.

However, prior to attempting to make a connection, we should ensure
that we have a (high enough) signal to make the connection. If not,
we now throw a SignalException.