CARS transaction adheres to server-side protocol
Ensure client can handle unexpected disconnect from server and
provide user with "useful" message.

Note, if the creation of an OutputStream continuously results
in an error and, thus throws an IOException, the result will
be that the client becomes stuck in a recurse loop. If this
occurs in practice we should remove the recursive call.